requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-modeltranslation==0.19.12
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
multidict==6.1.0
mysqlclient==2.2.7
openai==0.28.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pymemcache==4.0.0
PySocks==1.7.1
python-dotenv==1.0.1
python-memcached==1.62
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
Telethon==1.38.1
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
